REFACTOR #163 : like 패키지 구조 리팩토링
