+++ src/main/java/com/lokoko/domain/product/application/service/ProductReadService.java
@@ -3,9 +3,9 @@ package com.lokoko.domain.product.application.service;
 
 import com.lokoko.domain.image.entity.ProductImage;
 import com.lokoko.domain.image.repository.ProductImageRepository;
-import com.lokoko.domain.like.entity.ProductLike;
-import com.lokoko.domain.like.repository.ProductLikeRepository;
-import com.lokoko.domain.like.service.ProductLikeService;
+import com.lokoko.domain.like.domain.entity.ProductLike;
+import com.lokoko.domain.like.domain.repository.ProductLikeRepository;
+import com.lokoko.domain.like.application.service.ProductLikeService;
 import com.lokoko.domain.product.api.dto.NewProductProjection;
 import com.lokoko.domain.product.api.dto.PopularProductProjection;
 import com.lokoko.domain.product.api.dto.ReviewStats;
@@ -28,7 +28,6 @@ import com.lokoko.domain.product.domain.repository.ProductRepository;
 import com.lokoko.domain.product.exception.ProductNotFoundException;
 import com.lokoko.domain.product.mapper.ProductMapper;
 import com.lokoko.domain.review.api.dto.request.RatingCount;
-import com.lokoko.domain.review.domain.entity.enums.Rating;
 import com.lokoko.domain.review.domain.repository.ReviewRepository;
 import com.lokoko.global.common.response.PageableResponse;
 import com.lokoko.global.kuromoji.service.KuromojiService;
