+++ src/main/java/com/lokoko/domain/review/application/service/ReviewDetailsService.java
@@ -6,7 +6,7 @@ import com.lokoko.domain.image.entity.ReviewImage;
 import com.lokoko.domain.image.repository.ProductImageRepository;
 import com.lokoko.domain.image.repository.ReceiptImageRepository;
 import com.lokoko.domain.image.repository.ReviewImageRepository;
-import com.lokoko.domain.like.repository.ReviewLikeRepository;
+import com.lokoko.domain.like.domain.repository.ReviewLikeRepository;
 import com.lokoko.domain.product.domain.entity.Product;
 import com.lokoko.domain.review.api.dto.response.ImageReviewDetailResponse;
 import com.lokoko.domain.review.api.dto.response.VideoReviewDetailResponse;
